商品単価を小数点以下も扱えるように修正（但し、クライアント上では整数しか表示できない為、小数点以下の単価は0と表示される。）
git-svn-id: http://svn.l1j.org/l1j/trunk@555 698199e4-35c4-443b-82c4-7f4b4ae73c13

+++ db/schema/mysql/item_rates.sql
@@ -5,7 +5,7 @@ DROP TABLE IF EXISTS `item_rates`;
 CREATE TABLE `item_rates` (
   `item_id` int(10) unsigned NOT NULL,
   `note` varchar(255) DEFAULT NULL,
-  `selling_price` int(10) NOT NULL DEFAULT '-1',
-  `purchasing_price` int(10) NOT NULL DEFAULT '-1',
+  `selling_price` float NOT NULL DEFAULT '-1',
+  `purchasing_price` float NOT NULL DEFAULT '-1',
   PRIMARY KEY (`item_id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
